requirements/base.txt
cupy-cuda100==6.5.0
Faker==2.0.4
Flask-SocketIO==4.2.1
Flask==1.1.1
numba==0.46.0
numpy==1.17.3
Pillow==6.2.1
scipy==1.3.2
socketIO-client==0.7.2
